pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "askEIVA"
version = "0.1.2"
description = "AI Agent for the EIVA Support Team"
authors = [
    { name = "Philip Lau" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "mistralai>=0.0.12",
    "weaviate-client>=4.4.0",
    "unstructured>=0.12.0",    # For parsing the PDFs
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",        # For the Freshdesk API crawler
    "beautifulsoup4>=4.12.0",   # For HTML scraping if needed
    "pymupdf",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.285"
]

[tool.setuptools.packages.find]
where = ["src"]